+++ onlyoffice-wordpress.php
@@ -67,9 +67,10 @@ function deactivate_plugin_name() {
  * Uninstall hook.
  */
 function uninstall_onlyoffice_wordpress_plugin() {
-	delete_option( 'onlyoffice_settings' );
-	delete_option( 'onlyoffice-plugin-uuid' );
-	delete_option( 'onlyoffice-plugin-bytes' );
+	delete_site_option( 'onlyoffice_settings' );
+	delete_site_option( 'onlyoffice-plugin-uuid' );
+	delete_site_option( 'onlyoffice-plugin-bytes' );
+	delete_site_option( 'onlyoffice-formats' );
 }
 
 register_activation_hook( __FILE__, 'activate_plugin_name' );
@@ -108,7 +109,6 @@ run_plugin_name();
  * @return array
  */
 function onlyoffice_forms_mime_types( $mimes ) {
-	$mimes['oform'] = 'application/vnd.openxmlformats-officedocument.wordprocessingml.document.oform';
 	$mimes['docxf'] = 'application/vnd.openxmlformats-officedocument.wordprocessingml.document.docxf';
 
 	return $mimes;
@@ -128,10 +128,6 @@ function onlyoffice_forms_mime_types( $mimes ) {
 function onlyoffice_add_allow_upload_extension_exception( $data, $file, $filename ) {
 	$ext = pathinfo( $filename, PATHINFO_EXTENSION );
 	switch ( $ext ) {
-		case 'oform':
-			$data['ext']  = 'oform';
-			$data['type'] = 'application/vnd.openxmlformats-officedocument.wordprocessingml.document.oform';
-			break;
 		case 'docxf':
 			$data['ext']  = 'docxf';
 			$data['type'] = 'application/vnd.openxmlformats-officedocument.wordprocessingml.document.docxf';